+++ hazelcast-client-new/src/test/java/com/hazelcast/client/map/ClientMapLoaderExceptionHandlingTest.java
@@ -22,7 +22,7 @@ import java.util.HashSet;
 import java.util.Map;
 import java.util.Set;
 
-import static com.hazelcast.util.ValidationUtil.checkState;
+import static com.hazelcast.util.Preconditions.checkState;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
