+++ hazelcast/src/main/java/com/hazelcast/spi/impl/AbstractCompletableFuture.java
@@ -29,7 +29,7 @@ import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicReferenceFieldUpdater;
 
 import static com.hazelcast.util.ExceptionUtil.sneakyThrow;
-import static com.hazelcast.util.ValidationUtil.isNotNull;
+import static com.hazelcast.util.Preconditions.isNotNull;
 import static java.util.concurrent.atomic.AtomicReferenceFieldUpdater.newUpdater;
 
 public abstract class AbstractCompletableFuture<V> implements ICompletableFuture<V> {
